# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Visible groups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List groups visible to the current member",
        "tags": ["groups"]
      },
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Group created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Create a group and its owner membership atomically",
        "tags": ["groups"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/members": {
      "post": {
        "operationId": "GroupsController_addMember",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "A registered account was added to the group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredGroupMemberEnvelope"
                }
              }
            }
          },
          "202": {
            "description": "The account is not registered; a phone-bound registration and join invitation was sent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationSentEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Add a registered mobile account or send a secure group invitation",
        "tags": ["groups"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupMemberRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}": {
      "get": {
        "operationId": "GroupsController_detail",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get a group and its membership roster",
        "tags": ["groups"]
      }
    },
    "/api/v1/group-invitations/preview": {
      "post": {
        "operationId": "GroupInvitationsController_preview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Invitation details for the signed-in phone number.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationPreviewEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Preview a phone-bound group invitation after sign-in",
        "tags": ["group invitations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInvitationTokenRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/group-invitations/accept": {
      "post": {
        "operationId": "GroupInvitationsController_accept",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Invitation accepted and membership activated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInvitationAcceptanceEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Accept a phone-bound invitation and join its active group",
        "tags": ["group invitations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInvitationTokenRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "SPLITO API",
    "description": "Versioned API for SPLITO expense sharing and its financial journal",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current SPLITO deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SPLITO_SESSION"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
...
      "CreateGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "type": {
            "default": "other",
            "type": "string",
            "enum": ["home", "trip", "couple", "family", "project", "other"]
          },
          "defaultCurrency": {
            "default": "INR",
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "simplificationEnabled": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": ["name"],
        "additionalProperties": false
      },
      "AddGroupMemberRequest": {
        "type": "object",
        "properties": {
          "mobileNumber": {
            "type": "string",
            "minLength": 8,
            "maxLength": 40
          }
        },
        "required": ["mobileNumber"],
        "additionalProperties": false
      },
      "GroupInvitationTokenRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          }
        },
        "required": ["token"],
        "additionalProperties": false
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
                "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
              },
              "fieldErrors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["field", "message"],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Group": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "contextId",
          "name",
          "type",
          "defaultCurrency",
          "simplificationEnabled",
          "archived",
          "role",
          "memberCount",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "contextId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "example": "/api/v1/groups/id/image?v=media-id"
          },
          "type": {
            "type": "string",
            "enum": ["home", "trip", "couple", "family", "project", "other"]
          },
          "defaultCurrency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "INR"
          },
          "simplificationEnabled": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "administrator", "member"]
          },
          "memberCount": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GroupMember": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "kind", "role", "status", "allocationOrder"],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "example": "/api/v1/participants/id/avatar?v=media-id"
          },
          "kind": {
            "type": "string",
            "enum": ["USER", "GUEST"]
          },
          "role": {
            "type": "string",
            "enum": ["owner", "administrator", "member", "guest"]
          },
          "status": {
            "type": "string",
            "enum": ["active", "former"]
          },
          "allocationOrder": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "PendingGroupInvitation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "maskedMobileNumber", "expiresAt", "status"],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "maskedMobileNumber": {
            "type": "string",
            "example": "+********3210"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["pending"]
          }
        }
      },
      "RegisteredGroupMemberResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["outcome", "member"],
        "properties": {
          "outcome": {
            "type": "string",
            "enum": ["member_added"]
          },
          "member": {
            "$ref": "#/components/schemas/GroupMember"
          }
        }
      },
      "GroupInvitationSentResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["outcome", "invitation"],
        "properties": {
          "outcome": {
            "type": "string",
            "enum": ["invitation_sent"]
          },
          "invitation": {
            "$ref": "#/components/schemas/PendingGroupInvitation"
          },
          "developmentJoinUrl": {
            "type": "string",
            "format": "uri",
            "description": "Development-only captured join URL; never returned in production."
          }
        }
      },
      "GroupInvitationPreview": {
        "type": "object",
        "additionalProperties": false,
        "required": ["invitationId", "groupId", "groupName", "inviterDisplayName", "expiresAt"],
        "properties": {
          "invitationId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "groupId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "groupName": {
            "type": "string"
          },
          "inviterDisplayName": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GroupInvitationAcceptance": {
        "type": "object",
        "additionalProperties": false,
        "required": ["outcome", "groupId", "groupName", "member"],
        "properties": {
          "outcome": {
            "type": "string",
            "enum": ["joined"]
          },
          "groupId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "groupName": {
            "type": "string"
          },
          "member": {
            "$ref": "#/components/schemas/GroupMember"
          }
        }
      },
      "GroupDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Group"
          },
          {
            "type": "object",
            "required": ["members", "pendingInvitations"],
            "properties": {
              "members": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GroupMember"
                }
              },
              "pendingInvitations": {
                "type": "array",
                "description": "Pending mobile invitations, visible only to group owners and administrators.",
                "items": {
                  "$ref": "#/components/schemas/PendingGroupInvitation"
                }
              }
            }
          }
        ]
      },
...
      "GroupEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Group"
          }
        }
      },
      "GroupListEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Group"
            }
          }
        }
      },
      "GroupDetailEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupDetail"
          }
        }
      },
      "RegisteredGroupMemberEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegisteredGroupMemberResult"
          }
        }
      },
      "GroupInvitationSentEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupInvitationSentResult"
          }
        }
      },
      "GroupInvitationPreviewEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupInvitationPreview"
          }
        }
      },
      "GroupInvitationAcceptanceEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupInvitationAcceptance"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed input or identifier.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "A valid session is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated identity is not authorized.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not visible to this identity.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "The resource version, idempotency body, or financial state conflicts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied resource version is stale.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The uploaded file or multipart request exceeds its limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "The upload is not an accepted image or multipart body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The financial-domain invariant was not satisfied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required optimistic-concurrency header is missing.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The request rate limit was exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Minimum seconds before retrying; 60 for cooldown or up to 900 for quota.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 900
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BadGateway": {
        "description": "The external SMS provider did not accept the invitation message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required external delivery adapter is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}